Added support for expansions
It's now possible to define expansions for games. Expansions add more scoring categories to the base game, and optionally more misc fields.

When editing a play for a game that has expansions, the first step is to choose the used expansions. Selecting one or more expansions will cause the scoring and misc fields of the expansion to be added to the form after the scoring and misc fields of the base game (respectively).

Added the Leaders, Cities, and Armada expansions to 7 Wonders. Removed the old scoring category for leaders (as it's now included under the Leaders expansion), and the "Variant" misc fields for which expansions were used. Data migration for this is yet to come.