Fixed a bug in PlayForm related to field focusing
The field focusing logic in PlayForm was misbehaving with games that have expansions, causing the focusing logic to focus a field in the next view (instead of the current one). This was caused by an off-by-one error with games that have expansions. Before the expansion support, the currently active view index was always the same as the index of the currently filled field. The new expansion selection view causes these indices to be different (by one).

Separated the indexing of the active view from the indexing of the currently visible score/misc field to fix the problem.